+++ app/domain/model/taskModel/task_model_test.go
@@ -1,4 +1,4 @@
-package task_model
+package taskModel
 
 import (
 	"testing"
